changed the method from map to slice
The map didn't work as intended so I changed it to a slice. Now the only
thing in the instructions are the pointers. I also use a way to create
methods on the objects as well as created a `String()` method because it
was difficult to understand if my code did what it was intended to do.