Bug with removing
There is a problem when trying to see which instruction should be next
the preceding instructions are not checked if they have already been
fullfilled. This creates a wrong output. A way to solve this should be
to actually have a list with all preceding instructions. And if they are
done we add them to another queue. Then each instruction needs to be
checked if all the conditions have been met and only then we sort by the
lowest.